Replaces NLLLoss with CrossEntropyLoss - Removes log_softmax from architectures - Adds NLLLoss to override in pointer-generator - Removes  argument
